requirements.txt
faiss-cpu
insightface
ultralytics
django
onnxruntime
torch
torchvision
django-cors-headers
djangorestframework
